llに修正

+++ library/various/random.hpp
@@ -1,7 +1,7 @@
 #pragma once
-inline int random(int a, int b) {
+inline long long random(long long a, long long b) {
     if (a >= b) return a;
     static mt19937 mt(chrono::steady_clock::now().time_since_epoch().count());
-    uniform_int_distribution<int> dist(a, b - 1);
+    uniform_int_distribution<long long> dist(a, b - 1);
     return dist(mt);
 }
